+++ trench/command/generate_backup_codes.py
@@ -2,7 +2,7 @@ from django.utils.crypto import get_random_string
 
 from typing import Callable, Set
 
-from trench.settings import api_settings
+from trench.settings import trench_settings
 
 
 class GenerateBackupCodesCommand:
@@ -11,9 +11,9 @@ class GenerateBackupCodesCommand:
 
     def execute(
         self,
-        quantity: int = api_settings.BACKUP_CODES_QUANTITY,
-        length: int = api_settings.BACKUP_CODES_LENGTH,
-        allowed_chars: str = api_settings.BACKUP_CODES_CHARACTERS,
+        quantity: int = trench_settings.BACKUP_CODES_QUANTITY,
+        length: int = trench_settings.BACKUP_CODES_LENGTH,
+        allowed_chars: str = trench_settings.BACKUP_CODES_CHARACTERS,
     ) -> Set[str]:
         """
         Generates random encrypted backup codes.