+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * Type hits added
 * Backwards-compatible code refactoring and cleanup for better maintenance experience
 * Deactivation of primary MFA method would raise an error instead of randomly selecting new primary method from all active ones
+* Twilio backend client should be configured via environment variables instead of backend's configuration the `TWILIO_ACCOUNT_SID` and `TWILIO_AUTH_TOKEN`
 
 
 0.2.3 (2019-12-16)